pyproject.toml
[project]
name = "django-test-pj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.6",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "django-service-objects>=0.7.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "loguru>=0.7.3",
    "pre-commit>=4.3.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "rabbitmq-server>=0.0.1",
    "redis>=6.4.0",
    "ruff>=0.14.0",
    "setuptools>=80.9.0",
    "gunicorn>=20.1.0",
    "pillow>=12.0.0",
    "channels==4.1.0",
    "channels-redis==4.2.0",
    "daphne==4.1.0",
]
